A0-4231: Refactored all chain bootstraping code to separate module
