+++ bin/node/src/chain_spec/mod.rs
@@ -1,28 +1,25 @@
-use std::str::FromStr;
-
-use libp2p::PeerId;
-use primitives::AccountId;
-use serde::{de::Error, Deserialize, Deserializer, Serialize, Serializer};
-use sp_application_crypto::Pair;
-use sp_core::sr25519;
-
 mod builder;
 mod cli;
+pub mod commands;
+mod keystore;
 
-pub use cli::ChainParams;
+pub use commands::BootstrapChainCmd;
+pub use commands::ConvertChainspecToRawCmd;
 
 pub const CHAINTYPE_DEV: &str = "dev";
 pub const CHAINTYPE_LOCAL: &str = "local";
 pub const CHAINTYPE_LIVE: &str = "live";
 
 pub const DEFAULT_CHAIN_ID: &str = "a0dnet1";
-
-// Alice is the default sudo holder.
-pub const DEFAULT_SUDO_ACCOUNT: &str = "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY";
+pub const DEFAULT_BACKUP_FOLDER: &str = "backup-stash";
+pub const DEFAULT_SUDO_ACCOUNT_ALICE: &str = "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY";
 
 pub type AlephNodeChainSpec = sc_service::GenericChainSpec<()>;
 
-pub use cli::BootstrapChainCmd;
+use primitives::AccountId;
+use sc_chain_spec::ChainType;
+use sc_cli::Error;
+use sp_application_crypto::Ss58Codec;
 
 pub fn mainnet_config() -> Result<AlephNodeChainSpec, String> {
     AlephNodeChainSpec::from_json_bytes(crate::resources::mainnet_chainspec())
@@ -32,44 +29,16 @@ pub fn testnet_config() -> Result<AlephNodeChainSpec, String> {
     AlephNodeChainSpec::from_json_bytes(crate::resources::testnet_chainspec())
 }
 
-/// Generate an account ID from seed.
-pub fn account_id_from_string(seed: &str) -> AccountId {
-    AccountId::from(
-        sr25519::Pair::from_string(seed, None)
-            .expect("Can't create pair from seed value")
-            .public(),
-    )
+fn parse_chaintype(s: &str) -> Result<ChainType, Error> {
+    Ok(match s {
+        CHAINTYPE_DEV => ChainType::Development,
+        CHAINTYPE_LOCAL => ChainType::Local,
+        CHAINTYPE_LIVE => ChainType::Live,
+        s => panic!("Wrong chain type {s} Possible values: dev local live"),
+    })
 }
 
-#[derive(Clone)]
-pub struct SerializablePeerId {
-    inner: PeerId,
-}
-
-impl SerializablePeerId {
-    pub fn new(inner: PeerId) -> SerializablePeerId {
-        SerializablePeerId { inner }
-    }
-}
-
-impl Serialize for SerializablePeerId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let s: String = format!("{}", self.inner);
-        serializer.serialize_str(&s)
-    }
-}
-
-impl<'de> Deserialize<'de> for SerializablePeerId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        let inner = PeerId::from_str(&s)
-            .map_err(|_| D::Error::custom(format!("Could not deserialize as PeerId: {s}")))?;
-        Ok(SerializablePeerId { inner })
-    }
+/// Generate AccountId based on string command line argument.
+fn parse_account_id(s: &str) -> Result<AccountId, Error> {
+    Ok(AccountId::from_string(s).expect("Passed string is not a hex encoding of a public key"))
 }
