+++ bin/node/src/lib.rs
@@ -2,12 +2,10 @@ mod aleph_cli;
 mod aleph_node_rpc;
 mod chain_spec;
 mod cli;
-mod commands;
 mod executor;
 mod resources;
 mod rpc;
 mod service;
-mod shared_params;
 
 pub use cli::{Cli, Subcommand};
 #[cfg(any(
