+++ bin/node/src/service.rs
@@ -5,6 +5,7 @@ use std::{
     sync::Arc,
 };
 
+use crate::chain_spec::DEFAULT_BACKUP_FOLDER;
 use finality_aleph::{
     run_validator_node, AlephBlockImport, AlephConfig, AllBlockMetrics, BlockImporter,
     ChannelProvider, Justification, JustificationTranslator, MillisecsPerBlock, Protocol,
@@ -32,7 +33,6 @@ use crate::{
     aleph_cli::AlephCli,
     executor::aleph_executor,
     rpc::{create_full as create_full_rpc, FullDeps as RpcFullDeps},
-    shared_params::DEFAULT_BACKUP_FOLDER,
 };
 
 type AlephExecutor = aleph_executor::Executor;
